[refactor] 리뷰 내용 반영
