+++ src/main/java/com/prgrms/rg/domain/user/model/RiderProfile.java
@@ -10,6 +10,9 @@ import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.OneToMany;
 
+import com.prgrms.rg.domain.common.model.metadata.Bicycle;
+import com.prgrms.rg.domain.common.model.metadata.RidingLevel;
+
 import lombok.NoArgsConstructor;
 
 /**
@@ -28,7 +31,7 @@ public class RiderProfile {
 		서버 쪽에서는 그것을 시작 년도로 치환, 저장하여 사용자 편의성을 높이기로 하였음
 	 */
 	//TODO: User 생성자에서 UserBicycle 추가 불가능?
-	RiderProfile(int ridingYears, RidingLevel level) {
+	public RiderProfile(int ridingYears, RidingLevel level) {
 		this.ridingStartYear = Year.now().minusYears(ridingYears);
 		this.level = level;
 	}
